+++ packages/components/src/Button/Button.style.ts
@@ -22,26 +22,33 @@ export const useStyles = makeStyles(
 
       padding: theme.spacing(1.25, 2.5),
       fontSize: theme.spacing(2),
-      minWidth: theme.spacing(8.25),
       minHeight: theme.spacing(heightTabletXS),
       borderRadius: theme.spacing(heightTabletXS / 2),
+
+      '&$defaultMinWidth': {
+        minWidth: theme.spacing(8.25),
+      },
     },
 
     sizeExtraSmall: {
       padding: theme.spacing(0.5, 1),
       fontSize: theme.spacing(1.5),
-      minWidth: theme.spacing(13.25),
       minHeight: theme.spacing(extraSmallHeightTabletXS),
       borderRadius: theme.spacing(extraSmallHeightTabletXS / 2),
 
+      '&$defaultMinWidth': {
+        minWidth: theme.spacing(13.25),
+      },
+
       '&$outlinedPrimary': {
-        minWidth: theme.spacing(12.25),
+        '&$defaultMinWidth': {
+          minWidth: theme.spacing(12.25),
+        },
       },
     },
 
     sizeSmall: {
       fontSize: theme.spacing(2),
-      minWidth: theme.spacing(13.25),
       minHeight: theme.spacing(smallHeightTabletXS),
       borderRadius: theme.spacing(smallHeightTabletXS / 2),
 
@@ -50,17 +57,26 @@ export const useStyles = makeStyles(
         padding: '7px 20px',
       },
 
+      '&$defaultMinWidth': {
+        minWidth: theme.spacing(13.25),
+      },
+
       '&$outlinedPrimary': {
-        minWidth: theme.spacing(12.25),
+        '&$defaultMinWidth': {
+          minWidth: theme.spacing(12.25),
+        },
       },
     },
 
     sizeLarge: {
-      minWidth: theme.spacing(10),
       padding: theme.spacing(1.85, 4),
       fontSize: theme.spacing(2.5),
       minHeight: theme.spacing(largeHeightTabletXS),
       borderRadius: theme.spacing(largeHeightTabletXS / 2),
+
+      '&$defaultMinWidth': {
+        minWidth: theme.spacing(10),
+      },
     },
 
     outlinedPrimary: {
@@ -75,6 +91,12 @@ export const useStyles = makeStyles(
         background: `rgba(255, 255, 255, 0.2)`,
         color: `rgba(255, 255, 255, 0.5)`,
         border: 'none',
+
+        '&:hover:not(:active), &$focusVisible': {
+          background: `rgba(255, 255, 255, 0.2)`,
+          color: `rgba(255, 255, 255, 0.5)`,
+          border: 'none',
+        },
       },
 
       '&:before': {
@@ -129,6 +151,11 @@ export const useStyles = makeStyles(
       '&$disabled': {
         background: `rgba(255, 255, 255, 0.1)`,
         color: `rgba(255, 255, 255, 0.5)`,
+
+        '&:hover:not(:active), &$focusVisible': {
+          background: `rgba(255, 255, 255, 0.1)`,
+          color: `rgba(255, 255, 255, 0.5)`,
+        },
       },
 
       '&:before': {
@@ -190,6 +217,7 @@ export const useStyles = makeStyles(
 
     disabled: {},
     focusVisible: {},
+    defaultMinWidth: {},
 
     '@keyframes animate': {
       '0%': {