+++ packages/components/src/TabsSection/TabsSection.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import cn from 'classnames';
-import Tab, { TabTypeMap } from '@material-ui/core/Tab';
+import { TabTypeMap } from '@material-ui/core/Tab';
 import { TabListProps } from '@material-ui/lab';
 import TabContext from '@material-ui/lab/TabContext';
 import TabPanel from '@material-ui/lab/TabPanel';
 import { useBreakpointsMatch } from '@akropolis-web/styles';
 
+import { Tab } from '../Tab/Tab';
 import { TabList } from '../TabList/TabList';
 import { useStyles } from './TabsSection.style';
 
@@ -68,14 +69,7 @@ export function TabsSection<T extends React.ElementType = DefaultTabComponent>(p
   function renderTab(tabItem: TabItem<T>) {
     const { value, label, renderContent, ...restTabProps } = tabItem;
     return (
-      <Tab
-        {...restTabProps}
-        className={classes.tab}
-        label={label}
-        value={value}
-        key={value}
-        component={tabComponent}
-      />
+      <Tab {...restTabProps} label={label} value={value} key={value} component={tabComponent} />
     );
   }
 }