+++ packages/components/src/index.ts
@@ -9,9 +9,11 @@ export { default as IconButton } from '@material-ui/core/IconButton';
 export { default as Link, LinkProps } from '@material-ui/core/Link';
 export { default as Divider } from '@material-ui/core/Divider';
 export { default as Skeleton } from '@material-ui/lab/Skeleton';
+export { default as TabContext } from '@material-ui/lab/TabContext';
 
 export { Table };
 export { Button, ButtonProps } from './Button/Button';
+export { Tab, TabProps } from './Tab/Tab';
 export { TabList } from './TabList/TabList';
 export { TabsSection } from './TabsSection/TabsSection';
 export { Metric } from './Metric/Metric';