+++ packages/primitives/src/bnHexWei/format/balance/index.ts
@@ -1,0 +1,2 @@
+export { formatBalance } from './formatBalance';
+export { FormattedBalance } from './types';