+++ packages/primitives/src/entities/PercentAmount.ts
@@ -1,6 +1,6 @@
 import BN from 'bn.js';
 
-import { formatBalance } from '../bnHexWei';
+import { formatBalance, FormattedBalance } from '../bnHexWei';
 import { Value } from '../fraction';
 import { Amount } from './amount';
 import { Currency } from './Currency';
@@ -19,11 +19,16 @@ export class PercentAmount extends Amount<Currency> {
     return new PercentAmount(amount) as this;
   }
 
-  public toFormattedString(precision: number = 2, withSymbol = true): string {
+  public toFormattedBalance(
+    precision: number = 2,
+    withSymbol = true,
+    withSI?: boolean,
+  ): FormattedBalance {
     const multiplier = new BN(10).pow(new BN(precision));
     const value = this.toFraction().mul(multiplier);
 
     return formatBalance({
+      withSI,
       amountInBaseUnits: value.toBN(),
       tokenSymbol: withSymbol ? this.currency.symbol : undefined,
       baseDecimals: precision,