+++ packages/styles/src/theme.ts
@@ -9,8 +9,6 @@ import { getBreakpoints } from './getBreakpoints';
 
 export { Theme };
 
-const defaultTheme = createMuiTheme();
-
 function getGradients(type: 'dark' | 'light') {
   return {
     main: makeGradient([colors.heliotrope, colors.royalBlue]),
@@ -233,9 +231,12 @@ const sizes = {
     height: 36,
     indicatorSpace: 3,
     borderWidth: 1,
-    minWidth: 112,
   },
 };
+
+const defaultTheme = createMuiTheme({
+  breakpoints: getBreakpoints(breakpoints),
+});
 
 export const lightTheme = getTheme('light');
 export const darkTheme = getTheme('dark');
@@ -391,7 +392,7 @@ export function getTheme(type: 'light' | 'dark', overrides?: ThemeOptions): Them
           MuiOutlinedInput: {
             root: {
               borderColor: colors.darkMist,
-              borderRadius: 8,
+              borderRadius: 6,
               minHeight: 36,
 
               '&$focused': {
@@ -510,17 +511,20 @@ export function getTheme(type: 'light' | 'dark', overrides?: ThemeOptions): Them
 
           MuiTab: {
             root: {
-              minWidth: sizes.tabs.minWidth,
               position: 'relative',
               overflow: 'visible',
               minHeight: 'unset',
-              padding: defaultTheme.spacing(0.2, 1.5),
               textTransform: 'unset',
               fontSize: 'inherit',
               fontWeight: 300,
               lineHeight: 1.5,
               borderRadius:
                 sizes.tabs.height / 2 - sizes.tabs.indicatorSpace - sizes.tabs.borderWidth,
+              padding: defaultTheme.spacing(0.2, 1.25),
+
+              [defaultTheme.breakpoints.up('mobileMD')]: {
+                padding: defaultTheme.spacing(0.2, 2.5),
+              },
 
               '&::after': {
                 content: "''",
@@ -543,6 +547,27 @@ export function getTheme(type: 'light' | 'dark', overrides?: ThemeOptions): Them
               '&$selected': {
                 color: colors.white,
               },
+            },
+          },
+
+          MuiTooltip: {
+            tooltip: {
+              padding: '15px',
+              border: '1px solid #373740',
+              borderRadius: '6px',
+              backgroundColor: colors.blackRussian,
+            },
+            tooltipPlacementTop: {
+              margin: '8px 0 !important',
+            },
+            tooltipPlacementBottom: {
+              margin: '8px 0 !important',
+            },
+            tooltipPlacementRight: {
+              margin: '0 8px !important',
+            },
+            tooltipPlacementLeft: {
+              margin: '0 8px !important',
             },
           },
         },